Update python_daemon
python_daemon needs to be at >= v2.2.4 in order for this package to be
used with python 3.8. There is a known issue with using older versions
of python_daemon and python 3.8

Link: https://github.com/apache/airflow/issues/11456

As we are all moving towards updating our code to later versions of Python,
this change should hopefully be okay.