Add description column to variable (#286)
We started syncing Variable writes to both the Airflow 1 and 2
metastores. However, the Variable table slightly differs between the 2
versions. This introduces issues given the current way we write
variables to the Airflow 1 metastore from Airflow 2 as the column
"description" does not exist in Airflow 1.

I believe adding the column to the Airflow 1 metastore is the
simplest/cleanest fix.

An alternative could be to change how we write the variables from
Airflow 2 to 1 by taking the logic from Variables.set in 1.10.15 +
writing the sql itself and then adding that to lyft-etl util function.